pyproject.toml
[project]
name = "livraria"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "Vinyzada", email = "vinyescolas@gmail.com"},
]
dependencies = [
    "django>=5.0.3",
    "djangorestframework>=3.15.0",
    "django-cors-headers>=4.3.1",
]
requires-python = "==3.11.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false
